Clean manual list of packages.
With the recent packaging work, there's no more pip call at building or
installation time, and the extra dependencies that were added recently
can go away.

This would leave us with just adb, but it's already been added as a
dependency of the pirogue-tools package, itself pulled by pirogue-base,
so the file can stay empty (or might even be removed).